Revert "Move old parse tests to new test directory"
This reverts commit b52f0bc94aefd4b8a8a4f7927bb6537823255c08.

It turns out that most of the old parse tests simply aren't applicable to
parsing s-expressions and s-expression extensions. The syntax is so uniform with
so many fewer cases to consider that many of these test cases wind up being
duplicates or near duplicates of eachother when ported to the new syntax.